# cmd/api/docs/swagger.yaml
definitions:
  domain.Admin:
    properties:
      admin_name:
        maxLength: 12
        minLength: 4
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        maxLength: 15
        minLength: 8
        type: string
    required:
    - password
    type: object
  requests.Addcount:
    properties:
      count:
        minimum: 1
        type: integer
      product_id:
        type: integer
    required:
    - product_id
    type: object
...
  requests.BlockUser:
    properties:
      reason:
        type: string
      user_id:
        type: integer
    type: object
  requests.Cartreq:
    properties:
      productId:
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/block:
    patch:
      consumes:
      - application/json
      description: Admin can block a  user
      operationId: block-user
      parameters:
      - description: inputs
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.BlockUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can block a user
      tags:
      - Admin
...
  /admin/findall:
    get:
      consumes:
      - application/json
      operationId: FindAllUsers
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: FindAllUsers
      tags:
      - Admin
  /admin/finduser/{user_id}:
    get:
      consumes:
      - application/json
      description: Admin can fetch a specific user details using user id
      operationId: find-user-by-id
      parameters:
      - description: ID of the user to be fetched
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can fetch a specific user details using user id
      tags:
      - Admin
  /admin/login:
    post:
      consumes:
      - application/json
      operationId: LoginAdmin
      parameters:
      - description: Input Field
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: LoginAdmin
      tags:
      - Admin
  /admin/logout:
    post:
      description: Logout as a user exit from the ecommerce site
      operationId: AdminLogout
      responses:
        "200":
          description: success
        "400":
          description: failed
      summary: Adminlogout
      tags:
      - Admin
...
  /admin/salesreport:
    get:
      consumes:
      - application/json
      description: Admin can view the sales report
      operationId: view-sales-report
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can view sales report
      tags:
      - Admin
...
  /admin/signup:
    post:
      consumes:
      - application/json
      description: Save admin with details.
      operationId: SaveAdmin
      parameters:
      - description: Input Field
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/domain.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: SaveAdmin
      tags:
      - Admin
  /admin/unblock/{user_id}:
    patch:
      consumes:
      - application/json
      description: Admin can unblock a blocked user
      operationId: unblock-user
      parameters:
      - description: ID of the user to be unblocked
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can unblock a blocked user
      tags:
      - Admin
  /cart/AddToCart:
    post:
      description: user can add a stock in product to user cart
      operationId: AddToCart
      parameters:
      - description: Input true info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.Cartreq'
      responses:
        "200":
          description: Successfully productItem added to cart
        "400":
          description: can't add the product item into cart
      security:
      - ApiKeyAuth: []
      summary: api for add productItem to user cart
      tags:
      - Cart
  /cart/Addcount:
    put:
      consumes:
      - application/json
      description: user can delete their cartitems by id
      operationId: Add-Qantity
      parameters:
      - description: Input Field
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.Addcount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can delete a category
      tags:
      - Cart
  /cart/RemoveFromCart:
    delete:
      consumes:
      - application/json
      description: user can delete their cartitems by id
      operationId: delete-cartitem
      parameters:
      - description: Input Field
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.Cartreq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can delete a category
      tags:
      - Cart
  /cart/viewcart:
    get:
      description: user can see all productItem that stored in cart
      operationId: Cart
      responses:
        "200":
          description: successfully got user cart items
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: faild to get cart items
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: api for get all cart item of user
      tags:
      - Cart
...
swagger: "2.0"